AISMessage Insert

+++ AISapi/Models/AISMessage.cs
@@ -6,7 +6,7 @@ namespace AISapi.Models
         public int Id { get; set; }
         public DateTime Timestamp { get; set; }
         public string? Class { get; set; }
-        public long MMSI { get; set; }
+        public int MMSI { get; set; }
         public int? Vessel_IMO { get; set; }
         public MsgType Type { get; set; }
     }
